Sending and receiving UDP data

+++ GameServer/GameServer/Constants.cs
@@ -7,6 +7,6 @@ namespace GameServer
     class Constants
     {
         public const int TICKS_PER_SEC = 30;
-        public const int MS_PER_TICK = 1000 / TICKS_PER_SEC;
+        public const float MS_PER_TICK = 1000f / TICKS_PER_SEC;
     }
 }
